jira npoint resolve 9 Oct

+++ app/src/main/java/com/procialize/eventapp/MainActivity.java
@@ -70,6 +70,7 @@ import com.procialize.eventapp.ui.attendee.viewmodel.AttendeeViewModel;
 import com.procialize.eventapp.ui.eventList.view.EventListActivity;
 import com.procialize.eventapp.ui.eventinfo.view.EventInfoActivity;
 import com.procialize.eventapp.ui.home.view.HomeFragment;
+import com.procialize.eventapp.ui.livepoll.view.LivePollActivity;
 import com.procialize.eventapp.ui.login.view.LoginActivity;
 import com.procialize.eventapp.ui.newsfeed.view.NewsFeedFragment;
 import com.procialize.eventapp.ui.profile.model.Profile;
@@ -121,7 +122,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     ImageView headerlogoIv;
     RecyclerView rv_side_menu;
     boolean doubleBackToExitPressedOnce = false;
-    TableRow tr_switch_event, tr_home, tr_profile, tr_logout,tr_event_info;
+    TableRow tr_switch_event, tr_home, tr_profile, tr_logout,tr_event_info, tr_live_poll;
     TextView txt_version;
     LinearLayout ll_main;
     DatabaseReference mDatabaseReference;
@@ -235,6 +236,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         tr_event_info = findViewById(R.id.tr_event_info);
         tr_logout = findViewById(R.id.tr_logout);
         txt_version = findViewById(R.id.txt_version);
+        tr_live_poll = findViewById(R.id.tr_live_poll);
 
         txt_version.setText(BuildConfig.VERSION_NAME);
         tr_switch_event.setOnClickListener(this);
@@ -242,6 +244,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         tr_profile.setOnClickListener(this);
         tr_event_info.setOnClickListener(this);
         tr_logout.setOnClickListener(this);
+        tr_live_poll.setOnClickListener(this);
 
         if (tot_event.equalsIgnoreCase("1")) {
             tr_switch_event.setVisibility(View.GONE);
@@ -491,6 +494,11 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                         .replace(R.id.fragment_frame, EventInfoActivity.newInstance(), "")
                         .commit();*/
                 break;
+            case R.id.tr_live_poll:
+                /*JzvdStd.releaseAllVideos();
+                mDrawerLayout.closeDrawer(GravityCompat.START);
+                startActivity(new Intent(MainActivity.this, LivePollActivity.class));*/
+                break;
             case R.id.tr_logout:
                 JzvdStd.releaseAllVideos();
                 mDrawerLayout.closeDrawer(GravityCompat.START);
@@ -691,8 +699,10 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                             List<Attendee> attendeeList = event.getAttandeeList();
 
                             //Delete All attendee from local db and insert attendee
-                            attendeeDatabaseViewModel.deleteAllAttendee(MainActivity.this);
-                            attendeeDatabaseViewModel.insertIntoDb(MainActivity.this, attendeeList);
+                            if(attendeeList!=null) {
+                                attendeeDatabaseViewModel.deleteAllAttendee(MainActivity.this);
+                                attendeeDatabaseViewModel.insertIntoDb(MainActivity.this, attendeeList);
+                            }
                         }
 
                         if (attendeeViewModel != null && attendeeViewModel.getAttendeeList().hasObservers()) {
